Refactor JSP views and add interview feedback functionality
Replaced custom styles with Bootstrap for consistent UI across JSP views and improved layout for forms and lists. Added a new feature for recruiters to provide interview feedback, including creating `InterviewResult` enum, `InterviewFeedbackController`, and related JSP pages.

+++ src/main/java/com/jobrecruitment/repository/UserRepository.java
@@ -1,8 +1,6 @@
 package com.jobrecruitment.repository;
 
 import com.jobrecruitment.model.User;
-import com.jobrecruitment.model.company.Branch;
-import com.jobrecruitment.model.company.Company;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.data.jpa.repository.Query;
@@ -12,10 +10,10 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByEmail(String email);
 
     @Query("SELECT u FROM User u WHERE u.company.id = :companyId AND u.role = 'RECRUITER'")
-    List<User> findByCompanyIdAndRecruiterRole(@Param("companyId") Integer companyId);
+    List<User> findByCompanyIdAndRecruiterRole(@Param("companyId") Long companyId);
 }
